Fixed hyperlinks to local files
PDFs were not rendering due to bug in chromium, and markdowns opened locally were not compiling.
Instead pointed all of them at the soon-to-be-public COMPAS git repo, master branch.

This is a bit of a hack, since this is will only ever point at the master branch versions (which
currently don't even exist for some of them), but this should work broadly for the majority of cases.